Add basic unit test for storing number of visits for a particular class.

+++ src/main/java/com/google/sps/visits/VisitParent.java
@@ -5,9 +5,9 @@ import java.util.List;
 // Helper class to hold class name and visits lists
 public class VisitParent {
   private List<String> listOfClassNames;
-  private List<Integer> visitsPerClass;
+  private List<Long> visitsPerClass;
 
-  public VisitParent(List<String> names, List<Integer> visits) {
+  public VisitParent(List<String> names, List<Long> visits) {
     listOfClassNames = names;
     visitsPerClass = visits;
   }
